requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.9
click==8.0.1
cloudinary==1.28.0
Cython==0.29.24
dj-database-url==0.5.0
Django==3.2.5
django-cloudinary-storage==0.3.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
docx2txt==0.8
Flask==2.0.1
Flask-Cors==3.0.10
flask-paginate==0.8.1
Flask-PyMongo==2.3.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.2
Pillow==8.3.1
psycopg2==2.9.2
pymongo==3.12.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
Werkzeug==2.0.1